refactor: extract shared fetchChildrenByDepth helper; reuse PageOperations in FullPageViewOperations
- Extract the recursive children-fetching logic from BlockRetrievalOperations
  into src/tools/helpers/fetch-children.ts, used by both classes
- Remove duplicate fetchPageBlocks() (~60 lines) from FullPageViewOperations;
  call pageOps.fetchPageByUid() instead
- Remove duplicate fetchChildrenForBlocks() from FullPageViewOperations;
  call fetchChildrenByDepth() instead
- Pass pageOps from ToolHandlers constructor to FullPageViewOperations
