+++ src/tools/operations/block-retrieval.ts
@@ -2,6 +2,7 @@ import { Graph, q } from '@roam-research/roam-api-sdk';
 import { McpError, ErrorCode } from '@modelcontextprotocol/sdk/types.js';
 import { RoamBlock, RoamAncestor, RoamBlockWithAncestors } from '../../types/roam.js';
 import { resolveBlockRefs } from '../helpers/refs.js';
+import { fetchChildrenByDepth } from '../helpers/fetch-children.js';
 
 export class BlockRetrievalOperations {
   constructor(private graph: Graph) { }
@@ -12,51 +13,6 @@ export class BlockRetrievalOperations {
     }
 
     const block_uid = block_uid_raw.replace(/^\(\((.*)\)\)$/, '$1');
-
-    const fetchChildren = async (parentUids: string[], currentDepth: number): Promise<Record<string, RoamBlock[]>> => {
-      if (currentDepth >= depth || parentUids.length === 0) {
-        return {};
-      }
-
-      const childrenQuery = `[:find ?parentUid ?childUid ?childString ?childOrder ?childHeading
-                              :in $ [?parentUid ...]
-                              :where [?parent :block/uid ?parentUid]
-                                     [?parent :block/children ?child]
-                                     [?child :block/uid ?childUid]
-                                     [?child :block/string ?childString]
-                                     [?child :block/order ?childOrder]
-                                     [(get-else $ ?child :block/heading 0) ?childHeading]]`;
-
-      const childrenResults = await q(this.graph, childrenQuery, [parentUids]) as [string, string, string, number, number | null][];
-
-      const childrenByParent: Record<string, RoamBlock[]> = {};
-      const allChildUids: string[] = [];
-
-      for (const [parentUid, childUid, childString, childOrder, childHeading] of childrenResults) {
-        if (!childrenByParent[parentUid]) {
-          childrenByParent[parentUid] = [];
-        }
-        childrenByParent[parentUid].push({
-          uid: childUid,
-          string: childString,
-          order: childOrder,
-          heading: childHeading || undefined,
-          children: [],
-        });
-        allChildUids.push(childUid);
-      }
-
-      const grandChildren = await fetchChildren(allChildUids, currentDepth + 1);
-
-      for (const parentUid in childrenByParent) {
-        for (const child of childrenByParent[parentUid]) {
-          child.children = grandChildren[child.uid] || [];
-        }
-        childrenByParent[parentUid].sort((a, b) => a.order - b.order);
-      }
-
-      return childrenByParent;
-    };
 
     try {
       const rootBlockQuery = `[:find ?string ?order ?heading
@@ -72,7 +28,7 @@ export class BlockRetrievalOperations {
       }
 
       const [rootString, rootOrder, rootHeading] = rootBlockResults[0];
-      const childrenMap = await fetchChildren([block_uid], 0);
+      const childrenMap = await fetchChildrenByDepth(this.graph, [block_uid], depth);
 
       const rootBlock: RoamBlockWithAncestors = {
         uid: block_uid,
